up lại project đã chỉnh sửa phân quyền cho khuyến mãi vendor

+++ src/main/java/vn/iotstar/starshop/repository/VendorRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import vn.iotstar.starshop.entity.Vendor;
 
 @Repository
-public interface VendorRepository extends JpaRepository<Vendor, Long> {
+public interface VendorRepository extends JpaRepository<Vendor, Integer> {
     Vendor findByUserEmail(String email);  // Sử dụng email vì schema dùng email làm unique identifier
 }
